pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "RNTechScan"
version = "1.0.0"
description = "RNTechScan - Расширенный модульный сканер уязвимостей"
authors = [
    {name = "RNTechScan Team", email = "contact@rntechscan.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology", 
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: System :: Networking :: Monitoring",
]
dependencies = [
    "requests>=2.28.0",
    "PyYAML>=6.0",
]
keywords = ["vulnerability", "scanner", "security", "network", "web", "system"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0", 
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
]
enhanced = [
    "colorama>=0.4.5",
    "tqdm>=4.64.0",
    "click>=8.0.0",
]

[project.scripts]
rntechscan = "main:main"

[project.urls]
Homepage = "https://github.com/rntechscan/RNTechScan"
"Bug Reports" = "https://github.com/rntechscan/RNTechScan/issues"
"Source" = "https://github.com/rntechscan/RNTechScan"
"Documentation" = "https://rntechscan.readthedocs.io/"

[tool.setuptools]
packages = ["core", "modules", "reports", "config"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pyright]
include = [".", "core", "config", "modules"]
exclude = [".venv", "__pycache__", "*.pyc"]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
pythonPlatform = "Windows"
executeEnvironment = "virtualenv"

[tool.pyright.analysis]
autoSearchPaths = true
autoImportCompletions = true
diagnosticMode = "workspace"
typeCheckingMode = "basic"
useLibraryCodeForTypes = true
